Fixed PlayerInputs OnEnable
add to move to the PreInitializeMethod in order to avoid null reference and allow InputAction Enabling.

Refactored scene maangement with minor change to the button component that now solely handle AudioFX triggers instead of calling the function at 3 differents places. Still have to fix the issue where the buttons are being disable / interactive = false but arent being set back to true with the OnEnable function